Fixed problem with showing bad schedules
checkCombination() was fixed so now it uses itertools to find all the comparisons it should make to check if a schedule works. I probably changed some other stuff too, but I was on a different computer and didnt have my git stuff set up, so heres all the changes at once.